packages/python-packages/apiview-copilot/dev_requirements.txt
# regular requirements
aiohttp==3.13.3
azure-appconfiguration==1.7.1
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-keyvault-keys==4.11.0
azure-keyvault-secrets==4.10.0
azure-search-documents==11.5.2
black==25.1.0
colorama==0.4.6
fastapi==0.121.2
httpx==0.28.1
knack==0.12.0
matplotlib==3.10.0
openai==2.4.0
prompty[azure]==0.1.49
python-dotenv==1.0.1
pydantic==2.10.6
PyJWT==2.10.1
pyyaml==6.0.2
pytest==7.4.2
semantic-kernel==1.32.2
uvicorn[standard]==0.29.0

# evals requirements
azure-ai-evaluation==1.10.0
pytest==7.4.2
tabulate==0.9.0

# dev requirements
azure-ai-resources==1.0.0b9
azure-mgmt-appconfiguration==5.0.0
azure-mgmt-authorization==4.0.0
azure-mgmt-cognitiveservices==13.7.0
azure-mgmt-cosmosdb==9.8.0
azure-mgmt-keyvault==12.0.0
azure-mgmt-resource>=22.0.0,<23.0.0
azure-mgmt-search==9.1.0
azure-mgmt-web==8.0.0
beautifulsoup4==4.13.4
markdown-it-py==3.0.0
pylint==3.3.7
PyYAML==6.0.2
